Working version, now the feeds are being displayed after the addition.
Added a subject to InMemoryFeedRepository.kt to allow it to transmit event up the stream when the feeds list get changed. Every time it will send a new list upwords.
That led to changing the return type for GetAllFeedsUseCase to be Observable instead of just a Flowable, which actually makes sense now if you think about it. :)
Also added  executeObservable to BaseViewModel and used it in HomeViewModel.
Finally added ViewExtensions.kt with snack in it to be used instead of toast.